+++ data_provider/data_provider_factory.py
@@ -1,4 +1,5 @@
 from data_provider.base_data_provider import IDataProvider
+from data_provider.tinkoff_api_get_all_candles.tinkoff_getallcandles import TinkoffAPIGetAllCandles
 from data_provider.tinkoff_downloaded.tinkoff_downloaded import TinkoffDownloaded
 from data_provider.tinkoff_history_data.tinkoff_history_data import TinkoffHistoryData
 
@@ -17,5 +18,7 @@ class DataProviderFactory:
                 return TinkoffDownloaded(*args, **kwargs)
             case "TinkoffHistoryData":
                 return TinkoffHistoryData(*args, **kwargs)
+            case "TinkoffAPIGetAllCandles":
+                return TinkoffAPIGetAllCandles(*args, **kwargs)
             case _:
                 return TinkoffDownloaded(*args, **kwargs)