+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 2023-11-18
+### Added
+- 'TinkoffAPIGetAllCandles' data provider to get history candles from tinkoff api. 
+
 
 ## 2023-10-31
 ### Added